requirements.txt
Flask==2.1.0
Flask-Admin==1.5.8
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.5.2
Flask-Compress==1.11
Flask-HTTPAuth==4.2.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-mongoengine==0.9.5
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.9
Flask-Script==2.0.5
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
Flask-Testing==0.8.1
Flask-WTF==0.14.3
future==0.18.2
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==1.4.6
Flask-JWT==0.3.2
Werkzeug==2.0.0
PyYAML==6.0
Flask-JWT-Extended==4.4.4